+++ omc3/optics_measurements/measure_optics.py
@@ -97,7 +97,14 @@ def measure_optics(input_files: InputFiles, measure_input: DotDict) -> None:
             dispersion.calculate_normalised_dispersion(measure_input, input_files, beta_df, common_header)
 
     # Coupling ---
-    coupling.calculate_coupling(measure_input, input_files, phase_results, tune_dict, common_header)
+    try:
+        coupling.calculate_coupling(measure_input, input_files, phase_results, tune_dict, common_header)
+    except ZeroDivisionError:
+        if measure_input.accelerator.NAME in ('generic', 'sps'):
+            LOGGER.error("Coupling failed with zero-division error. Single plane BPMs? Skipped.")
+        else:
+            raise
+    
     if measure_input.only_coupling:
         return
     
@@ -132,7 +139,7 @@ def chromatic_beating(input_files: InputFiles, measure_input: DotDict, tune_dict
         betas = []
         for dpp_val in dpps:
             dpp_meas_input = deepcopy(measure_input)
-            dpp_meas_input["dpp"] = dpp_val
+            dpp_meas_input["analyse_dpp"] = dpp_val
             phase_res, out_dfs = phase.calculate(dpp_meas_input, input_files, tune_dict, plane)
             beta_df, _ = beta_from_phase.calculate(dpp_meas_input, tune_dict, phase_res[phase.COMPENSATED], {}, plane)
             betas.append(beta_df)